PWN-3266 - move same common code to BaseQrCodeFragment.kt
