+++ app/src/main/java/org/p2p/wallet/auth/ui/username/UsernameFragment.kt
@@ -1,6 +1,5 @@
 package org.p2p.wallet.auth.ui.username
 
-import android.Manifest
 import android.graphics.Bitmap
 import android.os.Bundle
 import android.view.View
@@ -8,15 +7,11 @@ import org.koin.android.ext.android.inject
 import org.p2p.wallet.R
 import org.p2p.wallet.auth.model.Username
 import org.p2p.wallet.common.analytics.interactor.ScreensAnalyticsInteractor
-import org.p2p.wallet.common.mvp.BaseMvpFragment
-import org.p2p.wallet.common.permissions.PermissionDeniedDialog
-import org.p2p.wallet.common.permissions.PermissionState
-import org.p2p.wallet.common.permissions.PermissionsDialog
-import org.p2p.wallet.common.permissions.PermissionsUtil
 import org.p2p.wallet.databinding.FragmentUsernameBinding
 import org.p2p.wallet.receive.analytics.ReceiveAnalytics
 import org.p2p.wallet.receive.list.TokenListFragment
-import org.p2p.wallet.receive.widget.QrView
+import org.p2p.wallet.receive.widget.BaseQrCodeFragment
+import org.p2p.wallet.receive.widget.ReceiveCardView
 import org.p2p.wallet.utils.Constants
 import org.p2p.wallet.utils.SpanUtils.highlightPublicKey
 import org.p2p.wallet.utils.popBackStack
@@ -24,13 +19,11 @@ import org.p2p.wallet.utils.replaceFragment
 import org.p2p.wallet.utils.shareScreenShot
 import org.p2p.wallet.utils.toast
 import org.p2p.wallet.utils.viewbinding.viewBinding
-import timber.log.Timber
 import java.io.File
 
 class UsernameFragment :
-    BaseMvpFragment<UsernameContract.View, UsernameContract.Presenter>(R.layout.fragment_username),
-    UsernameContract.View,
-    PermissionsDialog.Callback {
+    BaseQrCodeFragment<UsernameContract.View, UsernameContract.Presenter>(R.layout.fragment_username),
+    UsernameContract.View {
 
     companion object {
         fun create() = UsernameFragment()
@@ -99,43 +92,5 @@ class UsernameFragment :
         requireContext().shareScreenShot(qrImage, qrValue)
     }
 
-    override fun onPermissionsResult(state: Map<String, PermissionState>, payload: Any?) {
-        onCameraPermissionResult(state[Manifest.permission.WRITE_EXTERNAL_STORAGE])
-    }
-
-    private fun checkStatusAndRequestPermissionsIfNotGranted(): Boolean {
-        return PermissionsUtil.isGranted(
-            requireContext(),
-            Manifest.permission.WRITE_EXTERNAL_STORAGE
-        ).also { isGranted ->
-            if (!isGranted) {
-                PermissionsDialog.requestPermissions(
-                    this,
-                    listOf(Manifest.permission.WRITE_EXTERNAL_STORAGE)
-                )
-            }
-        }
-    }
-
-    private fun onCameraPermissionResult(state: PermissionState?) {
-        when (state) {
-            PermissionState.GRANTED -> {
-                with(binding.receiveCardView) {
-                    when (val action = getQrCodeLastAction()) {
-                        QrView.QrCodeAction.SHARE -> requestShare()
-                        QrView.QrCodeAction.SAVE -> requestSave()
-                        else -> {
-                            Timber.e("Unsupported QrCodeAction $action")
-                        }
-                    }
-                }
-            }
-            else -> PermissionDeniedDialog.show(
-                fragment = this,
-                permission = Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                title = getString(R.string.storage_permission_alert_title),
-                message = getString(R.string.storage_permission_alert_message)
-            )
-        }
-    }
+    override fun getReceiveCardView(): ReceiveCardView = binding.receiveCardView
 }
